moving structs into files
